+++ CHANGELOG.md
@@ -29,6 +29,8 @@ Given a version number `MAJOR.MINOR.PATCH`, we increment the:
   - `terramate experimental cloud login` -> `terramate cloud login`
   - `terramate experimental cloud info` -> `terramate cloud info`
   - `terramate experimental cloud drift show` -> `terramate cloud drift show`
+- Promote `--experimental-status` flag to `--cloud-status` flag
+  - `terramate experimental trigger --experimental-status=` -> `terramate experimental trigger --cloud-status=`
 
 ## 0.4.4